workaround sip error caused by sip upgrade (4.19.20->4.19.24)
Starting from 4.19.23, sip assumes that an assignment operator is available if a copy constructor is available. This is not the case for classes that contain const fields. As a
workaround, for each of these classes we declare a private copy assignment operator in the sip file (but not in the cpp header file, as the cpp compiler deletes the function by
default).

See also: https://www.riverbankcomputing.com/pipermail/pyqt/2020-June/043025.html